refactor: move variance-detail from components/ to features/reconciliation/
The VarianceDetail component and its types were misplaced under
src/components/reconciliation/ when they belong in
src/features/reconciliation/components/, consistent with the
feature-scoped directory convention used across the frontend.
