+++ frontend/src/features/reconciliation/components/variance-detail.tsx
